core/metamodel/src/test/java/org/nakedobjects/metamodel/facets/actions/exploration/ExplorationAnnotationFacetFactoryTest.java
package org.nakedobjects.metamodel.facets.actions.exploration;

import java.lang.reflect.Method;

import org.nakedobjects.applib.annotation.Exploration;
import org.nakedobjects.metamodel.facets.AbstractFacetFactoryTest;
import org.nakedobjects.metamodel.facets.Facet;
import org.nakedobjects.metamodel.spec.feature.NakedObjectFeatureType;


public class ExplorationAnnotationFacetFactoryTest extends AbstractFacetFactoryTest {

    private ExplorationAnnotationFacetFactory facetFactory;

    @Override
    protected void setUp() throws Exception {
        super.setUp();

        facetFactory = new ExplorationAnnotationFacetFactory();
    }

    @Override
    protected void tearDown() throws Exception {
        facetFactory = null;
        super.tearDown();
    }

    @Override
    public void testFeatureTypes() {
        final NakedObjectFeatureType[] featureTypes = facetFactory.getFeatureTypes();
        assertFalse(contains(featureTypes, NakedObjectFeatureType.OBJECT));
        assertFalse(contains(featureTypes, NakedObjectFeatureType.PROPERTY));
        assertFalse(contains(featureTypes, NakedObjectFeatureType.COLLECTION));
        assertTrue(contains(featureTypes, NakedObjectFeatureType.ACTION));
        assertFalse(contains(featureTypes, NakedObjectFeatureType.ACTION_PARAMETER));
    }

    public void testExplorationAnnotationPickedUp() {
        class Customer {
            @Exploration
            public void someAction() {}
        }
        final Method actionMethod = findMethod(Customer.class, "someAction");

        facetFactory.process(Customer.class, actionMethod, methodRemover, facetHolder);

        final Facet facet = facetHolder.getFacet(ExplorationFacet.class);
        assertNotNull(facet);
        assertTrue(facet instanceof ExplorationFacetAbstract);
        final ExplorationFacetAbstract executedFacetImpl = (ExplorationFacetAbstract) facet;

        assertNoMethodsRemoved();
    }

}

// Copyright (c) Naked Objects Group Ltd.
